Use size_t instead of unsigned and uint64_t

+++ frackcrypt/tools/base64/agent.cc
@@ -13,7 +13,7 @@ void Agent::reset() {
 }
 
 void Agent::chunkify(std::string input, unsigned chunk_size) {
-  for (uint64_t i = 0; i < input.length(); i += chunk_size)
+  for (size_t i = 0; i < input.length(); i += chunk_size)
     chunks.push_back(std::string(input, i, chunk_size));
 }
